Refactor and test RubyTimeParser
