+++ src/test/java/org/embulk/util/rubytime/TestRubyTimeParser.java
@@ -1,0 +1,25 @@
+package org.embulk.util.rubytime;
+
+import static org.junit.Assert.assertEquals;
+
+import java.time.temporal.TemporalAccessor;
+import org.junit.Test;
+
+/**
+ * Tests RubyTimeParser.
+ */
+public class TestRubyTimeParser {
+    @Test
+    public void testMultipleEpochs() {
+        final TemporalAccessor parsed1 = strptime("123456789 12849124", "%Q %s");
+        assertEquals(1000L * 12849124L, parsed1.getLong(RubyChronoField.Field.INSTANT_MILLIS));
+
+        final TemporalAccessor parsed2 = strptime("123456789 12849124", "%s %Q");
+        assertEquals(1000L * 3212281L / 250L, parsed2.getLong(RubyChronoField.Field.INSTANT_MILLIS));
+    }
+
+    private static TemporalAccessor strptime(final String string, final String format) {
+        final RubyTimeParser parser = new RubyTimeParser(RubyTimeFormat.compile(format));
+        return parser.parse(string);
+    }
+}
